- moved common unit test startup code into separate file
