+++ test/common/main-test.cpp
@@ -1,0 +1,38 @@
+/**
+ * This file just contains the startup of the unit test runners.
+ * The file can be used for x86 and arduino builds.
+*/
+#include <gtest/gtest.h>
+
+#if defined(ARDUINO)
+#include <Arduino.h>
+
+void setup()
+{
+    // should be the same value as for the `test_speed` option in "platformio.ini"
+    // default value is test_speed=115200
+    Serial.begin(115200);
+
+    ::testing::InitGoogleTest();
+}
+
+void loop()
+{
+	// Run tests
+	if (RUN_ALL_TESTS())
+	;
+
+	// sleep 1 sec
+	delay(1000);
+}
+
+#else
+int main(int argc, char **argv)
+{
+    ::testing::InitGoogleTest(&argc, argv);
+	if (RUN_ALL_TESTS())
+	;
+	// Always return zero-code and allow PlatformIO to parse results
+	return 0;
+}
+#endif
