+++ test/devices/test_jkbms/JkBmsTypesTest.cpp
@@ -61,7 +61,6 @@ TEST_F(JkBmsTypesTest, VerifyLowCapacityAlarmFlag)
 
 	JkBmsWarnMsg batWarnMsg(JKBMS_WARN_MSG_LOW_CAP_ALARM);
 
-
 	ASSERT_EQ(0b1, batWarnMsg.at<BatWarnMsgBits::LOW_CAP_ALARM>());
 }
 
@@ -156,38 +155,8 @@ TEST_F(JkBmsTypesTest, VerifyReservedBit15Flag)
 }
 
 } //namespace test
-
 } //namespace types
 
-#if defined(ARDUINO)
-#include <Arduino.h>
-
-void setup()
-{
-    // should be the same value as for the `test_speed` option in "platformio.ini"
-    // default value is test_speed=115200
-    Serial.begin(115200);
-
-    ::testing::InitGoogleTest();
-}
-
-void loop()
-{
-	// Run tests
-	if (RUN_ALL_TESTS())
-	;
-
-	// sleep 1 sec
-	delay(1000);
-}
-
-#else
-int main(int argc, char **argv)
-{
-    ::testing::InitGoogleTest(&argc, argv);
-	if (RUN_ALL_TESTS())
-	;
-	// Always return zero-code and allow PlatformIO to parse results
-	return 0;
-}
-#endif
+// Note: This is just a workaround, to prevent duplicate code for test application startup.
+//       If I have found a way to use multiple sources for unit tests within platformio, this include can be removed.
+#include <common/main-test.cpp>
